% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_createclass.R
\name{class_cleaner}
\alias{class_cleaner}
\title{Clean the Class}
\usage{
class_cleaner(hdir = "/home/shiny/DoubleAuction/Students/",
  classname = "ClassList_orig.csv", name_id = "Username",
  ID_id = "Student.ID", omits = NA)
}
\arguments{
\item{hdir}{directory holding master list of class}

\item{classname}{which class to include}

\item{name_id}{class names}

\item{name_id}{class ids}
}
\value{
nothing
}
\description{
Clean the Class
}
