% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_createclass.R
\name{class_sample}
\alias{class_sample}
\title{Format Class into Player Profiles}
\usage{
class_sample(seed = 33, ngame = 5, nstuds = 12,
  hdir = "/home/shiny/DoubleAuction/Students/", classname = "ClassList.csv")
}
\arguments{
\item{hdir}{directory holding master list of class}

\item{classname}{which class to include
#param cid which subset of players to include}
}
\value{
nothing
}
\description{
Format Class into Player Profiles
}
