% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_createclass.R
\name{create_class_fake}
\alias{create_class_fake}
\title{Create Fake Class}
\usage{
create_class_fake(n = 20, len = 8, patt = "[a-z]", ID = 1:n,
  gdir = NA)
}
\arguments{
\item{n}{number of students}

\item{len}{length of password}

\item{patt}{password pattern}

\item{ID}{student ids}

\item{gdir}{directory to save student list}
}
\value{
Class or name of saved file
}
\description{
Create Fake Class
}
