% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_groupmaker.R
\name{create_playergroups}
\alias{create_playergroups}
\title{Create Fake Class}
\usage{
create_playergroups(seed = 0, game_dir = NA, groupsize = 2, n = 20,
  len = 8, patt = "[a-z]", ID = 1:n)
}
\arguments{
\item{n}{number of students}

\item{len}{length of password}

\item{patt}{password pattern}

\item{ID}{student ids}

\item{gdir}{directory to save student list}
}
\value{
Class or name of saved file
}
\description{
Create Fake Class
}
