% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_startgame.R
\name{etime_fun}
\alias{etime_fun}
\title{Game Length Function Helper}
\usage{
etime_fun(H, M, S)
}
\arguments{
\item{H, M, S}{integers for hours,minutes,seconds}
}
\value{
Class or name of saved file
}
\description{
Game Length Function Helper
}
