% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_expdesign.R
\name{exp_design}
\alias{exp_design}
\title{Make Experiment}
\usage{
exp_design(buyers = c(1, 2, 12), sellers = 12, classtime = 75,
  sessiontime = 10)
}
\arguments{
\item{buyers, sellers}{number of students}

\item{classtime}{how long is the class}

\item{sessiontime}{how long will the games take}
}
\value{
nothing
}
\description{
Make Experiment
}
