% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_expdesign.R
\name{game_setup}
\alias{game_setup}
\title{Make Market Structure}
\usage{
game_setup(hdir = "~/DoubleAuction/Markets/", i, ngame = 5, ni = 12,
  gpass = FALSE)
}
\arguments{
\item{hdir}{home directory}

\item{i}{number of sellers}

\item{ngame}{number of games}

\item{ni}{number of buyers}

\item{gpass}{create participants passwords in gamefile}
}
\value{
nothing
}
\description{
Make Market Structure
}
