% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_startgame.R
\name{GameInit}
\alias{GameInit}
\title{Initiate Game}
\usage{
GameInit(TradePeriod = Sys.time(), startTime = Sys.time(),
  SessionName = "Trial", etime = SetupGameR::etime_fun("00", "10", "00"),
  gdir = "/srv/shiny-server/DoubleAuction/Game/", param_file = paste0(gdir,
  "time_file.RData"), u_timer = 60000, efreq = 1000, nperiods = 20,
  npractice = 2, ...)
}
\arguments{
\item{TradePeriod}{starting what trading period is the game?}

\item{startTime}{when to start game}

\item{gdir}{directory of game}

\item{u_timer, efreq}{timing parameters}

\item{nperiods, npractice}{lengths of periods}

\item{game_file}{name of game to initiate}
}
\value{
string, Name of Game File
}
\description{
Initiate Game
}
