% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_startgame.R
\name{GameLength}
\alias{GameLength}
\title{GameLength}
\usage{
GameLength(etime = etime_fun("00", "10", "00"), u_timer = 60000,
  gdir = "/srv/shiny-server/DoubleAuction/Game/", param_file = paste0(gdir,
  "time_file.RData"), ...)
}
\arguments{
\item{etime}{how long does the game endure?}

\item{gdir}{directory of game}

\item{param_file}{name of file with game parameters}
}
\value{
nothing
}
\description{
GameLength
}
