% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_stopgame.R
\name{GameStop}
\alias{GameStop}
\title{Stop A Game with a Pause}
\usage{
GameStop(game_file = paste0(gdir, "game_file.RData"), EXCEPT = c("gdir",
  "GameInit", "Game"), savepath = paste0(tempdir(), "/"),
  GameName = "Game_", RM = FALSE, SAVE = TRUE, period_rest = 5)
}
\arguments{
\item{game_file}{which game file to load}

\item{EXCEPT}{exceptions not to be removed}

\item{savepath}{file to save to}

\item{GameName}{name of file}

\item{RM}{remove objects in session?}

\item{SAVE}{save data}

\item{period_rest}{seconds to pause after shutdown}
}
\value{
nothing
}
\description{
Stop A Game with a Pause
}
