% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_playermaker.R
\name{IDsheet}
\alias{IDsheet}
\title{Create Profile List from Classlist}
\usage{
IDsheet(classlist, idsheetfun = idsheet_default, id = "ID", name = "Name",
  SIMPLIFY = FALSE, USE.NAMES = FALSE, ...)
}
\arguments{
\item{classlist}{matrix or data.frame indicating participants}

\item{idsheetfun}{which function to use}

\item{id}{column name with unique identifies}

\item{name}{column name with unique identifies}

\item{SIMPLIFY, USE.NAMES}{passed to mapply}

\item{...}{arguments passed to idsheetfun}
}
\value{
list of participants
}
\description{
Create Profile List from Classlist
}
