% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_playermaker.R
\name{idsheet_default}
\alias{idsheet_default}
\title{Format ID and Name into Empty Profile}
\usage{
idsheet_default(id_col, name_col)
}
\arguments{
\item{id_col}{Person ID}

\item{name_col}{Person Name}
}
\value{
a list, Empty userProfile
}
\description{
Format ID and Name into Empty Profile
}
