% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_playermaker.R
\name{player_maker}
\alias{player_maker}
\title{Format Class into Player Profiles}
\usage{
player_maker(gdir = NA, Class, IDsheetfun = IDsheet,
  idsheetfun = idsheet_default, id = "ID", name = "Name",
  appendw = "JA", SAVE = TRUE, ...)
}
\arguments{
\item{gdir}{directory to save player_file: Market/Game}

\item{Class}{which class to include}

\item{IDsheetfun}{which function to use}

\item{idsheetfun}{which function to use}

\item{id}{column name with unique identifies}

\item{name}{column name with unique identifies}

\item{appendw}{what to append student ID's with}

\item{SAVE}{save the participants}
}
\value{
Participants
}
\description{
Format Class into Player Profiles
}
