% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_periodtiming.R
\name{post_global}
\alias{post_global}
\title{Update Global Environment}
\usage{
post_global(...)
}
\arguments{
\item{...}{objects to post to the global environment}
}
\value{
Nothing
}
\description{
Update Global Environment
}
