% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_printfuns.R
\name{print_exists}
\alias{print_exists}
\alias{print_message}
\alias{print_proposals}
\alias{print_accepts}
\title{Prints Objects to Main Screen}
\usage{
print_exists(objname, fill = NA)

print_message(session)

print_proposals()

print_accepts()
}
\arguments{
\item{session}{session from server.R}
}
\value{
Nothing
}
\description{
Prints Objects to Main Screen
}
