% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_stopgame.R
\name{shutdown}
\alias{shutdown}
\title{Stop Game and Manage Output}
\usage{
shutdown(SAVE = TRUE, RM = FALSE, EXCEPT = NULL, startTime = Sys.time(),
  TradePeriod = Sys.time(), savepath = paste0(tempdir(), "/"),
  GameName = "Game_")
}
\arguments{
\item{SAVE}{save data}

\item{RM}{remove objects in session?}

\item{EXCEPT}{exceptions not to be removed}

\item{startTime}{what time did this game start at?}

\item{TradePeriod}{which trading period is this?}

\item{savepath}{file to save to}

\item{GameName}{name of file}
}
\value{
string, Name of Game File
}
\description{
Stop Game and Manage Output
}
