% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_printfuns.R
\name{sign_formatter2}
\alias{sign_formatter2}
\alias{sign_formatter}
\title{HTML formatting finance style}
\usage{
sign_formatter2(x, data = NULL)

sign_formatter(x, data = NULL)
}
\description{
HTML formatting finance style
}
\seealso{
sign_formatter

sign_formatter2
}
