% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_passwd.R
\name{sspasswd}
\alias{sspasswd}
\title{Create Password for Individual Player}
\usage{
sspasswd(user, passwd, create = FALSE, sys = TRUE,
  passfile = "/srv/shiny-server/passwd")
}
\arguments{
\item{user}{who to create a password for}

\item{passwd}{what password}

\item{create}{overwrite and create new passwd_file}

\item{sys}{if TRUE execute the bash script}

\item{passfile}{what file holds the passwords}
}
\value{
a string to be executed by system()
}
\description{
Create Password for Individual Player
}
