% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_passwd.R
\name{sspasswd_cmd}
\alias{sspasswd_cmd}
\title{Create Passwords for List of Players}
\usage{
sspasswd_cmd(Participants, passwd, create = FALSE, sys = FALSE,
  passfile = "/srv/shiny-server/passwd")
}
\arguments{
\item{Participants}{list of participants}

\item{passwd}{what password}

\item{create}{overwrite and create new passwd_file}

\item{sys}{if TRUE execute the bash script}

\item{passfile}{what file holds the passwords}
}
\value{
a string to be executed by system()
}
\description{
Create Passwords for List of Players
}
