% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Survey.R
\name{StartSurveyUI}
\alias{StartSurveyUI}
\alias{StartSurvey}
\title{Enter Survey Information}
\usage{
StartSurveyUI(id)

StartSurvey(input, output, session, GlobSurvey, SaveSurveyFile = getwd(),
  userPID)
}
\arguments{
\item{...}{}
}
\description{
Enter Survey Information
}
