% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_periodtiming.R
\name{stime_function}
\alias{stime_function}
\title{Check if current time is in a list of times}
\usage{
stime_function(stimes, e = 0)
}
\arguments{
\item{stimes}{list of times to check}

\item{e}{degree of rounding}
}
\value{
TRUE/FALSE
}
\description{
Check if current time is in a list of times
}
