% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_periodtiming.R
\name{timeLeft}
\alias{timeLeft}
\title{How much time left in trading session}
\usage{
timeLeft(startTime, etime, units = "mins")
}
\arguments{
\item{startTime}{a time object}

\item{etime}{a time object EndTime}

\item{units}{what units to return}
}
\value{
numeric object showing time in minutes
}
\description{
How much time left in trading session
}
