% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_userURLdata.R
\name{user_urldata_fun}
\alias{user_urldata_fun}
\title{Url Information About User}
\usage{
user_urldata_fun(user_data)
}
\arguments{
\item{user_data}{data from user}
}
\value{
string
}
\description{
Url Information About User
}
