% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_periodtiming.R
\name{utimeLeft}
\alias{utimeLeft}
\title{How much time left in trading session}
\usage{
utimeLeft(period_timer, tfreq = 1000, units = "secs")
}
\arguments{
\item{period_timer}{a time object}

\item{tfreq}{how often to update in milliseconds}
}
\value{
numeric object showing time in seconds
}
\description{
How much time left in trading session
}
