% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Survey.R
\name{ViewPaymentUI}
\alias{ViewPaymentUI}
\alias{ViewPayment}
\title{View Payment Information}
\usage{
ViewPaymentUI(id)

ViewPayment(input, output, session, GlobSurvey, userPID)
}
\arguments{
\item{...}{}
}
\description{
View Payment Information
}
