% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_startgame.R
\name{GameInit}
\alias{GameInit}
\title{Initiate Game}
\usage{
GameInit(TradePeriod = Sys.time(), startTime = Sys.time(),
  gdir = "/srv/shiny-server/DoubleAuction/Game/", game_file = paste0(gdir,
  "game_file.RData"))
}
\arguments{
\item{TradePeriod}{starting what trading period is the game?}

\item{startTime}{when to start game}

\item{gdir}{directory of game}

\item{game_file}{name of game to initiate}
}
\value{
string, Name of Game File
}
\description{
Initiate Game
}
